build: Introduce CROSS variable and use it to clean build tools setup.
When cross compiling in most of the cases the tool names are different only
by the cross compilation prefix added in front of the tool names. This patch
introduces CROSS variable to allows setting this prefix.

With this it is enough to run `make CROSS=/path/to/toolchain/sh4-linux- target`
or if tools are in the PATH `make CROSS=sh4-linux- target` and all tools will
be set. You can change both CROSS and DS_XX variables for example:
`make CROSS=/path/to/toolchain/osx-linux- DS_CC=clang target`