Make build output readable.
With this patch the build output shows what is going on but without
the details similar to Linux kernel build. This makes it easier to
see the warnings and errors. If you want to see the details set V
variable using make V=1 or export V=1 and then run make.