Fix compilation for targets that do not use libcrypto.
The build without libcrypto is already working in cmake build
but cscrypt/Makefile was missing rules for AES and SHA1 compilation.
This break the OSCAM build with messages about missing functions
when build without libcrypto was requested.

With this patch libcrypto is no longer needed for OSCAM compilation
except when WITH_SSL is defined.

git-svn-id: http://streamboard.gmc.to/svn/oscam/trunk@6666 4b0bc96b-bc66-0410-9d44-ebda105a78c1