Fixing tickets: 2592, 2531, 2574, 258
Ticket: 2592
thx to: theparasol and avs1978

We can use the other card frequencies of the internal reader of dm7025.
Possible impacts on other features

I think none.
What is the enhancement

Improves the code of theparasol for finding the divider of internal readers.

Further it finds the "divider" for the cardmhz setting of dm7025.

Use mhz=8300 for dm7025.

Ticket: 2574
thx to: Gompfs

Changed in globels.h mapprovid from unit16_t into uint32_t

Ticket: 2531
thx to: emm_dekoder

Fix infinite loop in viaccess.

Ticket: 2580
thx to: li-wii

Revision

all
Issue Description

Status not ok on VG2 reader(not due to kernel problem)
Explanation

When the card answer with a bad/or unknown statusbyte, the original STB reboot the card, why oscam doesn't?
Oscam sends ins also if card give bad answers, but we don't know if these ins can damage registers or anything else!
This patch do a full reboot of the card, not a fast reset because if there's an i/o error, oscam can stop the card. (with fast reset, card could go in reset loop)

I have made 2 check:
the first on status byte of d140
the second on status byte of d354

I don't think that card can give error only in one of the previous ins, and this is the safest way to preserve healty of cards.

With this patch you can delete all scripts that reboot oscam if a VG2 card gives "status not ok".

If you use a correct config (oscam.services, ecm whitelist, moderate overclock) and you keep the card costantly updated, with this patch you will forget of "status not ok"

..and yes, I have added some extra info on log to understand why sometimes card gives status ok, but cw 00 on subscribed channels



git-svn-id: http://streamboard.gmc.to/svn/oscam/trunk@7055 4b0bc96b-bc66-0410-9d44-ebda105a78c1